+++ utils/index.js
@@ -1,9 +1,11 @@
 const data = require('./data'),
     file = require('./file'),
+    directory = require('./directory'),
     profile = require('./profile');
 
 module.exports = {
     data,
     file,
+    directory,
     profile
 };
